refactor: hierarchical schema reorg
