+++ genson-core/src/tests/schema.rs
@@ -1,4 +1,4 @@
-// tests/schema.rs
+// genson-core/src/tests/schema.rs
 use super::*;
 use predicates::prelude::*;
 use serde_json::json;
